requirements.txt
torch
torchvision
numpy
scipy
pillow
tqdm
